+++ mne/beamformer/tests/test_lcmv.py
@@ -2,11 +2,11 @@ import os.path as op
 
 from nose.tools import assert_true
 import numpy as np
-# from numpy.testing import assert_array_almost_equal, assert_equal
+from numpy.testing import assert_array_almost_equal
 
 import mne
 from mne.datasets import sample
-from mne.beamformer import lcmv
+from mne.beamformer import lcmv, lcmv_epochs, lcmv_raw
 
 
 examples_folder = op.join(op.dirname(__file__), '..', '..', '..', 'examples')
@@ -28,11 +28,13 @@ label = mne.read_label(fname_label)
 noise_cov = mne.read_cov(fname_cov)
 raw = mne.fiff.Raw(fname_raw)
 forward = mne.read_forward_solution(fname_fwd)
+forward_fixed = mne.read_forward_solution(fname_fwd, force_fixed=True,
+                                          surf_ori=True)
 events = mne.read_events(fname_event)
 
 
 def test_lcmv():
-    """Test LCMV
+    """Test LCMV with evoked data and single trials
     """
     event_id, tmin, tmax = 1, -0.2, 0.2
 
@@ -41,8 +43,10 @@ def test_lcmv():
 
     # Set up pick list: EEG + MEG - bad channels (modify to your needs)
     left_temporal_channels = mne.read_selection('Left-temporal')
-    picks = mne.fiff.pick_types(raw.info, meg=True, eeg=False, stim=True, eog=True,
-                       exclude=raw.info['bads'], selection=left_temporal_channels)
+    picks = mne.fiff.pick_types(raw.info, meg=True, eeg=False,
+                                stim=True, eog=True,
+                                exclude=raw.info['bads'],
+                                selection=left_temporal_channels)
 
     # Read epochs
     epochs = mne.Epochs(raw, events, event_id, tmin, tmax, proj=True,
@@ -64,3 +68,60 @@ def test_lcmv():
 
     assert_true(0.09 < tmax < 0.1)
     assert_true(2. < np.max(max_stc) < 3.)
+
+    # Now test single trial using fixed orientation forward solution
+    # so we can compare it to the evoked solution
+    stcs = lcmv_epochs(epochs, forward_fixed, noise_cov, data_cov, reg=0.01)
+
+    epochs.drop_bad_epochs()
+    assert_true(len(epochs.events) == len(stcs))
+
+    # average the single trial estimates
+    stc_avg = np.zeros_like(stc.data)
+    for this_stc in stcs:
+        stc_avg += this_stc.data
+    stc_avg /= len(stcs)
+
+    # compare it to the solution using evoked with fixed orientation
+    stc_fixed = lcmv(evoked, forward_fixed, noise_cov, data_cov, reg=0.01)
+    assert_array_almost_equal(stc_avg, stc_fixed.data)
+
+
+def test_lcmv_raw():
+    """Test LCMV with raw data
+    """
+    tmin, tmax = 0, 20
+    # Setup for reading the raw data
+    raw.info['bads'] = ['MEG 2443', 'EEG 053']  # 2 bads channels
+
+    # Set up pick list: EEG + MEG - bad channels (modify to your needs)
+    left_temporal_channels = mne.read_selection('Left-temporal')
+    picks = mne.fiff.pick_types(raw.info, meg=True, eeg=False, stim=True,
+                                eog=True, exclude=raw.info['bads'],
+                                selection=left_temporal_channels)
+
+    noise_cov = mne.read_cov(fname_cov)
+    noise_cov = mne.cov.regularize(noise_cov, raw.info,
+                                   mag=0.05, grad=0.05, eeg=0.1, proj=True)
+
+    start, stop = raw.time_to_index(tmin, tmax)
+
+    # use only the left-temporal MEG channels for LCMV
+    picks = mne.fiff.pick_types(raw.info, meg=True, exclude=raw.info['bads'],
+                                selection=left_temporal_channels)
+
+    data_cov = mne.compute_raw_data_covariance(raw, tmin=tmin, tmax=tmax)
+
+    stc = lcmv_raw(raw, forward, noise_cov, data_cov, reg=0.01, label=label,
+                   start=start, stop=stop, picks=picks)
+
+    assert_array_almost_equal(np.array([tmin, tmax]),
+                              np.array([stc.times[0], stc.times[-1]]),
+                              decimal=2)
+
+    # make sure we get an stc with vertices only in the lh
+    vertno = [forward['src'][0]['vertno'], forward['src'][1]['vertno']]
+    assert_true(len(stc.vertno[0]) == len(np.intersect1d(vertno[0],
+                                                         label['vertices'])))
+    assert_true(len(stc.vertno[1]) == 0)
+    # TODO: test more things